+++ newton-4.00/sdk/dBrain/ndBrainFloatBuffer.cpp
@@ -137,7 +137,7 @@ void ndBrainFloatBuffer::VectorFromDevice(ndBrainVector& vector) const
 void ndBrainFloatBuffer::VectorToDevice(const ndBrainVector& vector)
 {
 	ndAssert(vector.GetCount() <= ndInt64(m_sizeInBytes / sizeof(ndReal)));
-	MemoryToDevice(0, vector.GetCount() * sizeof(ndReal), &vector[0]);
+	MemoryToDevice(0, size_t(vector.GetCount() * sizeof(ndReal)), &vector[0]);
 }
 
 void ndBrainFloatBuffer::Set(const ndBrainFloatBuffer& buffer)
